+++ src/contextimpl.rs
@@ -43,13 +43,6 @@ where
     handles: SmallVec<[SpawnHandle; 2]>,
 }
 
-unsafe impl<A> Send for ContextImpl<A>
-where
-    A: Actor + Send,
-    A::Context: AsyncContext<A>,
-{
-}
-
 impl<A> ContextImpl<A>
 where
     A: Actor,