+++ src/lib.rs
@@ -46,7 +46,6 @@ extern crate tokio_executor;
 extern crate tokio_io;
 extern crate tokio_reactor;
 extern crate tokio_tcp;
-extern crate tokio_threadpool;
 extern crate tokio_timer;
 
 #[macro_use]